requirements.txt
discord.py
wavelink
yt-dlp
